+++ changelog.md
@@ -109,3 +109,5 @@
 - TODO save dialog
 
 - TODO update language to match fegs-dashboard.xlsm
+
+- TODO error when focus leaves txtExpln